requirements.txt
Django>=4.2,<5.0
celery>=5.2,<6.0
redis>=4.5,<5.0
django-celery-beat>=2.6,<3.0
daphne>=4.1,<5.0
django-import-export>=3.0,<4.0
channels>=4.1.0
gunicorn>=20.2.4 


